// pharmacy-management-frontend/src/app/model/prescriptionDto.ts
import {PrescriptionIndicative} from './prescription-indicative';
import {Indicative} from './indicative';

export interface PrescriptionDto {
  prescriptionId?: number;
  prescriptionCode?: string;
  prescriptionName?: string;
  symptom?: string;
  object?: string;
  flag?: boolean;
  numberOfDay?: number;
  note?: string;
  indicatives: Indicative[];
}